# api/openapi.yaml
openapi: 3.1.0
info:
  title: bili-notify administration API
  version: 3.0.0
  description: |
    Authoritative transport contract for the same-origin TLS administration UI.
    Version 3 unifies Bilibili and Knowledge Planet accounts, sources, content and comment trees.
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/v3/accounts/bilibili/qr:
    get:
      tags: [accounts]
      operationId: getBilibiliQRLogin
      responses:
        '200':
          description: Current Bilibili QR login transaction or null.
          content:
            application/json:
              schema:
                oneOf:
                  - {$ref: '#/components/schemas/BilibiliLogin'}
                  - {type: 'null'}
    post:
      tags: [accounts]
      operationId: startBilibiliQRLogin
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '201':
          description: Bilibili QR login transaction.
          content: {application/json: {schema: {$ref: '#/components/schemas/BilibiliLogin'}}}
  /api/v3/accounts/bilibili/qr/{id}:
    delete:
      tags: [accounts]
      operationId: cancelBilibiliQRLogin
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '204': {description: QR login transaction cancelled.}
  /api/v3/accounts/bilibili/session:
    delete:
      tags: [accounts]
      operationId: deleteBilibiliSession
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses: {'204': {description: Bilibili account disconnected.}}
  /api/v3/accounts/zsxq/sms-code:
    post:
      tags: [accounts]
      operationId: sendZSXQSMSCode
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ZSXQSMSCodeRequest'}
      responses:
        '201':
          description: Ephemeral login transaction.
          content: {application/json: {schema: {$ref: '#/components/schemas/ZSXQLoginTransaction'}}}
        '429': {$ref: '#/components/responses/RateLimitError'}
  /api/v3/accounts/zsxq/session:
    post:
      tags: [accounts]
      operationId: createZSXQSession
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [transaction_id, code]
              properties: {transaction_id: {type: string}, code: {type: string}}
      responses:
        '201':
          description: Connected Knowledge Planet account.
          content: {application/json: {schema: {$ref: '#/components/schemas/PlatformAccount'}}}
    delete:
      tags: [accounts]
      operationId: deleteZSXQSession
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses: {'204': {description: Knowledge Planet account disconnected.}}
  /api/v3/accounts/zsxq/sync-sources:
    post:
      tags: [accounts, sources]
      operationId: syncZSXQSources
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '200':
          description: Visible Knowledge Planet sources.
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Source'}}}}
  /api/v3/sources:
    get:
      tags: [sources]
      operationId: listSources
      parameters:
        - {name: platform, in: query, schema: {type: string, enum: [bilibili, zsxq]}}
      responses:
        '200':
          description: Collection sources.
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Source'}}}}
    post:
      tags: [sources]
      operationId: createSource
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateSourceRequest'}}}
      responses:
        '201': {description: Source created, content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}}
  /api/v3/sources/{id}:
    parameters: [{name: id, in: path, required: true, schema: {type: string}}]
    put:
      tags: [sources]
      operationId: updateSource
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateSourceRequest'}}}
      responses: {'200': {description: Source updated, content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}}}
    delete:
      tags: [sources]
      operationId: deleteSource
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses: {'204': {description: Source and archive removed.}}
  /api/v3/contents:
    get:
      tags: [history]
      operationId: listContents
      parameters:
        - {name: platform, in: query, schema: {type: string, enum: [bilibili, zsxq]}}
        - {name: source_id, in: query, schema: {type: string}}
        - {name: q, in: query, schema: {type: string}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
        - {name: after, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        '200':
          description: Unified content page.
          content: {application/json: {schema: {$ref: '#/components/schemas/ContentPage'}}}
  /api/v3/contents/{id}:
    get:
      tags: [history]
      operationId: getContent
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses:
        '200':
          description: Content with attachment metadata.
          content:
            application/json:
              schema:
                type: object
                required: [content, attachments]
                properties:
                  content: {$ref: '#/components/schemas/Content'}
                  attachments: {type: array, items: {$ref: '#/components/schemas/Attachment'}}
  /api/v3/contents/{id}/comments:
    get:
      tags: [history]
      operationId: getContentCommentTree
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses:
        '200':
          description: Stable nested comment tree.
          content:
            application/json:
              schema:
                type: object
                required: [children, incomplete]
                properties:
                  children: {type: array, items: {$ref: '#/components/schemas/CommentTreeNode'}}
                  incomplete: {type: boolean}
  /api/v3/contents/{id}/attachments/{attachment_id}:
    get:
      tags: [history]
      operationId: downloadAttachment
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: attachment_id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Authenticated attachment stream with Range support.}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bili_notify_session
  headers:
    RequestID:
      description: Server-generated request correlation identifier.
      schema:
        type: string
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 1}
...
  responses:
...
    RateLimitError:
      description: Authentication attempts are rate limited.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
...
    PlatformAccount:
      type: object
      additionalProperties: false
      required: [platform, status]
      properties:
        platform: {type: string, enum: [bilibili, zsxq]}
        external_id: {type: string}
        display_name: {type: string}
        masked_phone: {type: string}
        status: {type: string, enum: [disconnected, connected, invalid, risk_paused]}
        verified_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        risk_paused_until: {type: string, format: date-time}
        last_error: {type: string}
    ZSXQSMSCodeRequest:
      type: object
      additionalProperties: false
      required: [country_code, phone, captcha_verify_param, agreement_accepted]
      properties:
        country_code: {type: string, pattern: '^\+[0-9]{1,4}$'}
        phone: {type: string, pattern: '^[0-9]{5,20}$'}
        captcha_verify_param: {type: string, minLength: 1}
        agreement_accepted: {type: boolean}
    ZSXQLoginTransaction:
      type: object
      additionalProperties: false
      required: [id, masked_phone, expires_at, next_send_at, attempts_left]
      properties:
        id: {type: string}
        masked_phone: {type: string}
        expires_at: {type: string, format: date-time}
        next_send_at: {type: string, format: date-time}
        attempts_left: {type: integer, minimum: 0, maximum: 5}
    Source:
      type: object
      additionalProperties: false
      required: [id, platform, type, external_id, name, enabled, baseline_state, backfill_done, backfill_total, sync_lag_sec, consecutive_fails]
      properties:
        id: {type: string}
        platform: {type: string, enum: [bilibili, zsxq]}
        type: {type: string, enum: [up, planet]}
        external_id: {type: string}
        name: {type: string}
        note: {type: string}
        owner_id: {type: string}
        owner_name: {type: string}
        enabled: {type: boolean}
        baseline_state: {type: string, enum: [pending, running, complete, failed]}
        backfill_cursor: {type: string}
        high_watermark: {type: string}
        backfill_done: {type: integer, format: int64}
        backfill_total: {type: integer, format: int64}
        last_poll_at: {type: string, format: date-time}
        last_success_at: {type: string, format: date-time}
        last_comment_at: {type: string, format: date-time}
        sync_lag_sec: {type: integer, format: int64}
        last_error: {type: string}
        consecutive_fails: {type: integer}
    CreateSourceRequest:
      type: object
      additionalProperties: false
      required: [platform, external_id, name, note, enabled]
      properties:
        platform: {type: string, const: bilibili}
        external_id: {type: string, pattern: '^[1-9][0-9]*$'}
        name: {type: string}
        note: {type: string}
        enabled: {type: boolean}
    UpdateSourceRequest:
      type: object
      additionalProperties: false
      required: [name, note, enabled]
      properties:
        name: {type: string}
        note: {type: string}
        enabled: {type: boolean}
    Content:
      type: object
      additionalProperties: false
      required: [id, platform, source_id, external_id, upstream_type, type, published_at, first_seen_at, last_synced_at, baseline]
      properties:
        id: {type: string}
        platform: {type: string, enum: [bilibili, zsxq]}
        source_id: {type: string}
        external_id: {type: string}
        author_id: {type: string}
        author_name: {type: string}
        upstream_type: {type: string}
        type: {type: string, enum: [dynamic, video, article, discussion, question, answer, task, long_article]}
        title: {type: string}
        text: {type: string}
        safe_html: {type: string}
        url: {type: string}
        published_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        first_seen_at: {type: string, format: date-time}
        last_synced_at: {type: string, format: date-time}
        deleted_at: {type: string, format: date-time}
        stats: {type: object, additionalProperties: {type: integer, format: int64}}
        tree_incomplete: {type: boolean}
        baseline: {type: boolean}
    Attachment:
      type: object
      additionalProperties: false
      required: [id, content_id, external_id, type]
      properties:
        id: {type: string}
        content_id: {type: string}
        external_id: {type: string}
        type: {type: string, enum: [image, file, audio, video, link]}
        file_name: {type: string}
        mime: {type: string}
        size: {type: integer, format: int64}
        width: {type: integer}
        height: {type: integer}
        duration_sec: {type: integer, format: int64}
        remote_host: {type: string}
        local_path: {type: string}
        localize_error: {type: string}
    CommentTreeNode:
      type: object
      additionalProperties: false
      required: [id, platform, content_id, rpid, mid, name, message, time]
      properties:
        id: {type: string}
        platform: {type: string, enum: [bilibili, zsxq]}
        content_id: {type: string}
        root_id: {type: string}
        parent_id: {type: string}
        author_id: {type: string}
        author_role: {type: string, enum: [owner, admin, guest, partner, member, up]}
        updated_at: {type: string, format: date-time}
        first_seen_at: {type: string, format: date-time}
        deleted_at: {type: string, format: date-time}
        media: {type: array, items: {$ref: '#/components/schemas/Attachment'}}
        children: {type: array, items: {$ref: '#/components/schemas/CommentTreeNode'}}
        rpid: {type: string}
        parent: {type: string}
        mid: {type: string}
        name: {type: string}
        message: {type: string}
        time: {type: string, format: date-time}
        is_up: {type: boolean}
        is_trigger: {type: boolean}
    ContentPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Content'}}
        page: {$ref: '#/components/schemas/PageMetadata'}
...
    PageMetadata:
      type: object
      additionalProperties: false
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: string
          description: Empty when has_more is false; otherwise pass unchanged as after.
        has_more: {type: boolean}
...
    BilibiliLogin:
      type: object
      additionalProperties: false
      required: [id, status, expires_at]
      properties:
        id: {type: string}
        status: {type: string}
        expires_at: {type: string, format: date-time}
        qr_data_url: {type: string}
...